+++ src/main/java/com/shreyas/url_shortner/url/UrlRepository.java
@@ -3,7 +3,7 @@ package com.shreyas.url_shortner.url;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UrlRepository extends JpaRepository<URL, Integer> {
+public interface UrlRepository extends JpaRepository<URL, Long> {
     // Custom query to find a URL by its short code
     Optional<URL> findByShortCode(String shortCode);
 }
